refactor: 메서드 네이밍과 코드포멧 통일, conflict 해결
