+++ src/main/java/knu/fest/knu/fest/domain/booth/service/BoothService.java
@@ -1,30 +1,37 @@
 package knu.fest.knu.fest.domain.booth.service;
 
 import knu.fest.knu.fest.domain.booth.controller.dto.BoothCreateRequest;
+import knu.fest.knu.fest.domain.booth.controller.dto.BoothCreateResponse;
 import knu.fest.knu.fest.domain.booth.controller.dto.BoothDetailResponse;
 import knu.fest.knu.fest.domain.booth.controller.dto.BoothUpdateRequest;
 import knu.fest.knu.fest.domain.booth.entity.Booth;
 import knu.fest.knu.fest.domain.booth.repository.BoothRepository;
+import knu.fest.knu.fest.domain.comment.controller.dto.CommentResponse;
+import knu.fest.knu.fest.domain.comment.repository.CommentRepository;
 import knu.fest.knu.fest.global.exception.CommonException;
 import knu.fest.knu.fest.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class BoothService {
 
     private final BoothRepository boothRepository;
+    private final CommentRepository commentRepository;
 
     @Transactional
-    public BoothDetailResponse createBooth(BoothCreateRequest request) {
+    public BoothCreateResponse createBooth(BoothCreateRequest request) {
         if (boothRepository.existsByBoothNumber(request.boothNumber())) {
             throw new CommonException(ErrorCode.ALREADY_EXIST_BOOTH_NUMBER);
         }
         Booth booth = request.toEntity();
+        Booth saved = boothRepository.save(booth);
 
-        return BoothDetailResponse.of(boothRepository.save(booth));
+        return BoothCreateResponse.from(saved);
     }
 
     @Transactional
@@ -33,8 +40,12 @@ public class BoothService {
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_BOOTH));
 
         booth.update(request.name(), request.description());
+        List<CommentResponse> comments = commentRepository.findAllByBoothId(id)
+                .stream()
+                .map(CommentResponse::from)
+                .toList();
 
-        return BoothDetailResponse.of(booth);
+        return BoothDetailResponse.of(booth, comments);
     }
 
     @Transactional(readOnly = true)
@@ -42,7 +53,12 @@ public class BoothService {
         Booth booth = boothRepository.findById(id)
                 .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_BOOTH));
 
-        return BoothDetailResponse.of(booth);
+        List<CommentResponse> comments = commentRepository.findAllByBoothId(id)
+                .stream()
+                .map(CommentResponse::from)
+                .toList();
+
+        return BoothDetailResponse.of(booth, comments);
     }
 
 /**
