+++ src/main/java/knu/fest/knu/fest/domain/comment/entity/Comment.java
@@ -36,4 +36,8 @@ public class Comment extends BaseEntity {
         this.user = user;
         this.content = content;
     }
+
+    public boolean isOwner(Long userId) {
+        return user.getId().equals(userId);
+    }
 }
