+++ src/main/java/knu/fest/knu/fest/global/exception/ErrorCode.java
@@ -15,6 +15,7 @@ public enum ErrorCode {
     NOT_FOUND_LIKE(40403, HttpStatus.NOT_FOUND, "존재하지 않는 좋아요입니다."),
     NOT_FOUND_PLACE(40403, HttpStatus.NOT_FOUND, "존재하지 않는 건물입니다."),
     NOT_FOUND_BOOTH(40404, HttpStatus.NOT_FOUND, "존재하지 않는 부스입니다."),
+    NOT_FOUND_COMMENT(40405, HttpStatus.NOT_FOUND, "존재하지 않는 댓글입니다."),
     NOT_FOUND_WAITING(40405, HttpStatus.NOT_FOUND, "존재하지 않는 웨이팅입니다."),
 
 
@@ -31,6 +32,7 @@ public enum ErrorCode {
     TOKEN_EXPIRED(40301, HttpStatus.FORBIDDEN, "토큰이 만료되었습니다."),
     INVALID_JWT_TOKEN_FORMAT(40302, HttpStatus.FORBIDDEN, "토큰 형식이 맞지 않습니다. 올바른 토큰 형식 'Authorization' : 'Bearer {token}')"),
     POST_EDIT_FORBIDDEN(40303, HttpStatus.FORBIDDEN, "게시물을 수정할 권한이 없습니다."),
+    COMMENT_EDIT_FORBIDDEN(40304, HttpStatus.FORBIDDEN, "댓글을 수정할 권한이 없습니다."),
 
     // Internal Server Error
     INTERNAL_SERVER_ERROR(50000, HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류입니다.")
